rebased 264: fix warnings with -Wall

+++ motoman_driver/src/industrial_robot_client/joint_relay_handler.cpp
@@ -297,7 +297,7 @@ bool JointRelayHandler::select(const JointTrajectoryPoint& all_joint_state,
   pub_joint_names->clear();
 
   // skip over "blank" joint names
-  for (int i = 0; i < all_joint_names.size(); ++i)
+  for (size_t i = 0; i < all_joint_names.size(); ++i)
   {
     if (all_joint_names[i].empty())
       continue;
@@ -325,7 +325,7 @@ bool JointRelayHandler::select(const DynamicJointsGroup& all_joint_state,
   pub_joint_names->clear();
 
   // skip over "blank" joint names
-  for (int i = 0; i < all_joint_names.size(); ++i)
+  for (size_t i = 0; i < all_joint_names.size(); ++i)
   {
     if (all_joint_names[i].empty())
       continue;
